Add files via upload
Currently, you would have to change the file path to test it out as I took the local path of the xlsx file. So, change the path to test the code.
NOTE: Only EnhancedAlgo3.js contains the concept of atoms but the rest 2 codes are probably giving the same output even without using atoms.